requirements.txt

python-fasthtml